seed.go の内容を app.goに組み込んだ
